add more tests for evaluation of reduced gradient
